+++ bin/confluence.sh
@@ -3,34 +3,16 @@ set -e
 set -x
 set -u
 
-## See more docs at https://hub.docker.com/r/atlassian/confluence-server/
+MY_TMP_CONTEXT="${HOME}/docker-files/confluence"
 
-
-MY_DOCKER_IMAGE=atlassian/confluence-server:6.7.2
-MY_TMP_CONTEXT="tmpcontext-confluence-6.7"
-
-[ ! -d ${MY_TMP_CONTEXT} ] && mkdir ${MY_TMP_CONTEXT}
+[ ! -d ${MY_TMP_CONTEXT} ] && exit 1
 cd ${MY_TMP_CONTEXT}
-cat > docker-compose.yml <<EOF
-version: "3.5"
-
-services:
-  myconfluence:
-    container_name: myconfluence
-    image: ${MY_DOCKER_IMAGE}
-    ports:
-      - 127.0.0.1:8090:8090
-      - 127.0.0.1:8091:8091
-    restart: always
-    volumes:
-      - confluence_data:/var/atlassian/application-data/confluence
-
-volumes:
-  confluence_data:
-EOF
 
 set +u
 case ${1} in
+    open)
+        open -a Firefox http://127.0.0.1:8090
+        ;;
     stop)
         ## Destroy the stack but keep the data
         docker-compose down
@@ -38,6 +20,15 @@ case ${1} in
     destroy)
         ## Destroy the stack and data
         docker-compose down --volumes
+        ;;
+    logs)
+        docker-compose logs
+        ;;
+    ps)
+        docker-compose ps
+        ;;
+    top)
+        docker-compose top
         ;;
     start|*)
         docker-compose pull
@@ -49,4 +40,3 @@ case ${1} in
 esac
 
 cd -
-rm -r ${MY_TMP_CONTEXT}
