+++ bin/jenkins-lts.sh
@@ -3,34 +3,16 @@ set -e
 set -x
 set -u
 
-## See more docs at https://github.com/jenkinsci/docker/blob/master/README.md
+MY_TMP_CONTEXT="${HOME}/docker-files/jenkins-lts"
 
-# MY_DOCKER_IMAGE=jenkins/jenkins:lts
-MY_DOCKER_IMAGE=jenkins/jenkins:2.107.1
-MY_TMP_CONTEXT="tmpcontext-jenkins-2.107"
-
-[ ! -d ${MY_TMP_CONTEXT} ] && mkdir ${MY_TMP_CONTEXT}
+[ ! -d ${MY_TMP_CONTEXT} ] && exit 1
 cd ${MY_TMP_CONTEXT}
-cat > docker-compose.yml <<EOF
-version: "3.5"
-
-services:
-  myjenkins:
-    container_name: myjenkins
-    image: ${MY_DOCKER_IMAGE}
-    ports:
-      - 127.0.0.1:8080:8080
-      - 127.0.0.1:50000:50000
-    restart: always
-    volumes:
-      - jenkins_home:/var/jenkins_home
-
-volumes:
-  jenkins_home:
-EOF
 
 set +u
 case ${1} in
+    open)
+        open -a Firefox http://127.0.0.1:8080
+        ;;
     stop)
         ## Destroy the stack but keep the data
         docker-compose down
@@ -38,6 +20,15 @@ case ${1} in
     destroy)
         ## Destroy the stack and data
         docker-compose down --volumes
+        ;;
+    logs)
+        docker-compose logs
+        ;;
+    ps)
+        docker-compose ps
+        ;;
+    top)
+        docker-compose top
         ;;
     start|*)
         docker-compose pull
@@ -49,4 +40,3 @@ case ${1} in
 esac
 
 cd -
-rm -r ${MY_TMP_CONTEXT}
